fix: make index optional

+++ tools/picard.wdl
@@ -318,12 +318,14 @@ task sort {
             --CREATE_MD5_FILE true \
             --VALIDATION_STRINGENCY ~{validation_stringency}
 
-        mv ~{prefix}.bai ~{outfile_name}.bai
+        if [ -f "~{prefix}.bai" ]; then
+            mv ~{prefix}.bai ~{outfile_name}.bai
+        fi
     >>>
 
     output {
         File sorted_bam = outfile_name
-        File sorted_bam_index = outfile_name + ".bai"
+        File? sorted_bam_index = outfile_name + ".bai"
         File sorted_bam_md5 = outfile_name + ".md5"
     }
